[WIP] bumping to rust 1.0

+++ ttt/console_ui/runner.rs
@@ -20,7 +20,7 @@ fn run_game(board: &mut Board, player_1: Player, player_2: Player) {
         presenter::present_board(board);
 
         let (current_player, opponent): (Player, Player) = rules::current_player_and_opponent(board.cells, player_1, player_2);
-        let chosen_cell: uint = current_player.choose_next_move(board, (current_player.token, opponent.token));
+        let chosen_cell: usize = current_player.choose_next_move(board, (current_player.token, opponent.token));
         let cells = board.cells.as_mut_slice();
         let token = current_player.token;
